DICOMStore/requirements.txt
pydicom>=2.4.0
pynetdicom>=2.0.0
PyYAML>=6.0
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
